requirements.txt
Pyrogram==2.0.26
python-dotenv==0.20.0